+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Next
 - Changed output of `getVersions` to replace '`.x`' with the actual game version.
+- Changed CLI output of `--list` to present the versions in prose instead of JSON.
 
 ## 1.3.16
 - Updated resource pack format to `36`.