+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 1.3.16
+*2024-08-24*
 - Updated resource pack format to `36`.
 - Updated data pack format to `50`.
 - Fixed special case checking not being applied to full version names leading to no result for certain versions.