+++ changelog.md
@@ -4,6 +4,7 @@
 - Added basic message when no arguments are given.
 - Changed output of `getVersions` to replace '`.x`' with the actual game version.
 - Changed CLI output of `--list` to present the versions in prose instead of JSON.
+- Changed help message.
 
 ## 1.3.16
 - Updated resource pack format to `36`.